+++ restocker-deliveries/src/main/java/de/restockoffice/article/ArticleClient.java
@@ -1,4 +1,4 @@
-package de.restockoffice;
+package de.restockoffice.article;
 
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.Path;
