+++ restocker-deliveries/src/main/java/de/restockoffice/article/ArticleDto.java
@@ -1,4 +1,4 @@
-package de.restockoffice;
+package de.restockoffice.article;
 
 public class ArticleDto {
     private String productId;
