+++ restocker-deliveries/src/main/java/de/restockoffice/delivery/CustomerDeliveryOverviewDto.java
@@ -1,4 +1,4 @@
-package de.restockoffice;
+package de.restockoffice.delivery;
 
 public class CustomerDeliveryOverviewDto {
     private DeliverySummaryDto lastDelivery;
