+++ restocker-deliveries/src/main/java/de/restockoffice/delivery/DeliveredArticleSummaryDto.java
@@ -1,4 +1,4 @@
-package de.restockoffice;
+package de.restockoffice.delivery;
 
 public class DeliveredArticleSummaryDto {
