+++ restocker-deliveries/src/main/java/de/restockoffice/delivery/DeliveryPublicationDto.java
@@ -1,4 +1,4 @@
-package de.restockoffice;
+package de.restockoffice.delivery;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
