+++ restocker-deliveries/src/main/java/de/restockoffice/delivery/MonthlyDeliveryCustomersDto.java
@@ -1,4 +1,4 @@
-package de.restockoffice;
+package de.restockoffice.delivery;
 
 import java.util.List;
