+++ restocker-deliveries/src/main/java/de/restockoffice/order/OrderClient.java
@@ -1,4 +1,4 @@
-package de.restockoffice;
+package de.restockoffice.order;
 
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.HeaderParam;
