+++ restocker-deliveries/src/main/java/de/restockoffice/order/OrderDto.java
@@ -1,4 +1,4 @@
-package de.restockoffice;
+package de.restockoffice.order;
 
 import java.time.LocalDateTime;
