+++ restocker-deliveries/src/main/java/de/restockoffice/user/UserClient.java
@@ -1,4 +1,4 @@
-package de.restockoffice;
+package de.restockoffice.user;
 
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.HeaderParam;
